Use stricter encoding anywhere we might pass UGC
With the new built-in feature to import arbitrary content packs, its a
good idea to tighten up encoding since we are inserting content directly
into the tetrio.js file.

This json parses base64 encoded data instead of evalling it, which was
mostly completely safe beforehand since the user would insert their own
content. I say 'mostly' because you could type arbitrary strings into
the music metadata, but that's just as difficult as convincing someone
to type stuff into the developer console.

Either way this fixes a currently non-existant XSS exploit since the
data is already JSON.stringifed and inserted into known safe locations,
but hey, defense in depth. The only way the previous version could be
exploited is if tetrio's source changed in a way that made the filters
insert the json somewhere where it wouldn't be parsed as json (for
example in the middle of a string with doublequotes). Such a change
would be highly unlikely in the first place but far more likely to just
break the filters instead.

Note that missing-music-patch-filter.js, osd-request-filter.js, and
april-fools-filter.js are not included because we're inserting
hardcoded (and therefore trusted) content.